requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=2.120=mkl
blas-devel=3.9.0=20_win64_mkl
brotli-python=1.1.0=py311h12c1d0e_1
bzip2=1.0.8=hcfcfb64_5
ca-certificates=2023.11.17=h56e8100_0
certifi=2023.11.17=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=win_pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_0
cuda-cccl=12.3.101=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-opencl-dev=12.3.101=0
cuda-profiler-api=12.3.101=0
cuda-runtime=12.1.0=0
cudatoolkit=11.5.0=hfde6d1b_9
eigen=3.4.0=h91493d7_0
ffmpeg=4.2.3=ha925a31_0
filelock=3.13.1=pyhd8ed1ab_0
freetype=2.12.1=hdaf720e_2
gettext=0.21.1=h5728263_0
glib=2.78.1=h12be248_1
glib-tools=2.78.1=h12be248_1
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.12.1=nompi_h57737ce_104
icu=58.2=ha925a31_3
idna=3.5=pyhd8ed1ab_0
intel-openmp=2023.2.0=h57928b3_50497
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=h8ffe710_2
kornia=0.7.0=pyhd8ed1ab_0
krb5=1.21.2=heb0366b_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=20_win64_mkl
libcblas=3.9.0=20_win64_mkl
libclang=12.0.0=default_h627e005_2
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.4.101=0
libcurand-dev=10.3.4.101=0
libcurl=8.4.0=hd5e4a3a_0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libexpat=2.5.0=h63175ca_1
libffi=3.4.2=h8ffe710_5
libglib=2.78.1=h16e383f_1
libiconv=1.17=h8ffe710_0
libjpeg-turbo=2.0.0=h196d8e1_0
liblapack=3.9.0=20_win64_mkl
liblapacke=3.9.0=20_win64_mkl
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h19919ed_0
libprotobuf=3.20.3=h12be248_0
libsqlite=3.44.2=hcfcfb64_0
libssh2=1.11.0=h7dfc565_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=hcfcfb64_1
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.2=hcfcfb64_1
libwebp-base=1.3.2=hcfcfb64_0
libxcb=1.15=hcd874cb_0
libxml2=2.10.4=hc3477c8_0
libxslt=1.1.37=h0192164_0
libzlib=1.2.13=hcfcfb64_5
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.3=py311ha68e1ae_1
mkl=2023.2.0=h6a75c08_50497
mkl-devel=2023.2.0=h57928b3_50497
mkl-include=2023.2.0=h6a75c08_50497
mpmath=1.3.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
networkx=3.2.1=pyhd8ed1ab_0
ninja=1.11.1=h91493d7_0
numpy=1.26.2=py311h0b4df5a_0
opencv=4.6.0=py311h5d08a89_5
opencv-python=4.8.1.78=pypi_0
openssl=3.2.0=hcfcfb64_1
packaging=23.2=pyhd8ed1ab_0
pcre2=10.42=h17e33f8_0
pillow=9.4.0=py311hd77b12b_1
pip=23.3.1=pyhd8ed1ab_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pysocks=1.7.1=pyh0701188_6
python=3.11.6=h2628c8c_0_cpython
python_abi=3.11=4_cp311
pytorch=2.1.1=py3.11_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py311ha68e1ae_1
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
requests=2.31.0=pyhd8ed1ab_0
scipy=1.11.4=py311h0b4df5a_0
setuptools=68.2.2=pyhd8ed1ab_0
sqlite=3.44.2=hcfcfb64_0
sympy=1.12=pyh04b8f61_3
tbb=2021.8.0=h59b6b97_0
tk=8.6.13=h5226925_1
torchaudio=2.1.1=pypi_0
torchvision=0.16.1=pypi_0
tqdm=4.66.1=pyhd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.1.0=pyhd8ed1ab_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
wheel=0.41.3=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=h12be248_0
